MongodbFastAPI/requirements.txt
pymongo
fastapi
uvicorn
python-dotenv
